feat: put npm versioning logic in a separate file so that it can be used as a standalone command
